No longer bubble tag exceptions to the top
When a tag has an invalid component to it it would cause a whole file to fail
to parse. Instead we should consider invalid tags a nuisance but not worth
breaking the whole parsing over.

In this change I have added a try..catch that will prevent invalid tags from
breaking the whole parsing process.